Refactor ResponsiveStory from middleware to stories
